+++ class.rb
@@ -1,0 +1,11 @@
+require_relative 'module'
+class MyList
+  include MyEnumerable
+  def initialize(*args)
+    @list = args
+  end
+
+  def each(&block)
+    @list.each(&block)
+  end
+end
